Homework 4/services/scrapers/issuer_news/requirements.txt
setuptools
aiohttp
aiolimiter
asyncpg
requests
beautifulsoup4
lxml
PyPDF2
python-dotenv